Refactor composables, split into smaller files
