+++ feature/pairalert/src/main/java/dev/arkbuilders/rate/feature/pairalert/presentation/main/PairAlertConditionScreen.kt
@@ -2,23 +2,14 @@
 
 package dev.arkbuilders.rate.feature.pairalert.presentation.main
 
-import android.Manifest
 import android.content.Context
 import android.widget.Toast
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.pager.PagerState
@@ -31,27 +22,17 @@ import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHostState
-import androidx.compose.material3.Switch
-import androidx.compose.material3.SwitchDefaults
 import androidx.compose.material3.Text
 import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.annotation.ExternalModuleGraph
@@ -59,15 +40,11 @@ import com.ramcosta.composedestinations.generated.pairalert.destinations.AddPair
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import com.ramcosta.composedestinations.result.ResultRecipient
 import com.ramcosta.composedestinations.result.onResult
-import dev.arkbuilders.rate.core.domain.CurrUtils
-import dev.arkbuilders.rate.core.presentation.CoreRDrawable
 import dev.arkbuilders.rate.core.presentation.CoreRString
 import dev.arkbuilders.rate.core.presentation.theme.ArkColor
-import dev.arkbuilders.rate.core.presentation.ui.AppButton
 import dev.arkbuilders.rate.core.presentation.ui.AppHorDiv16
 import dev.arkbuilders.rate.core.presentation.ui.AppSwipeToDismiss
 import dev.arkbuilders.rate.core.presentation.ui.AppTopBarCenterTitle
-import dev.arkbuilders.rate.core.presentation.ui.CurrIcon
 import dev.arkbuilders.rate.core.presentation.ui.GroupViewPager
 import dev.arkbuilders.rate.core.presentation.ui.LoadingScreen
 import dev.arkbuilders.rate.core.presentation.ui.NotifyAddedSnackbarVisuals
@@ -77,14 +54,12 @@ import dev.arkbuilders.rate.core.presentation.ui.group.EditGroupOptionsBottomShe
 import dev.arkbuilders.rate.core.presentation.ui.group.EditGroupRenameBottomSheet
 import dev.arkbuilders.rate.core.presentation.ui.group.EditGroupReorderBottomSheet
 import dev.arkbuilders.rate.core.presentation.ui.group.EditGroupRow
-import dev.arkbuilders.rate.core.presentation.utils.DateFormatUtils
 import dev.arkbuilders.rate.feature.pairalert.di.PairAlertComponent
 import dev.arkbuilders.rate.feature.pairalert.di.PairAlertComponentHolder
 import dev.arkbuilders.rate.feature.pairalert.domain.model.PairAlert
 import kotlinx.coroutines.launch
 import org.orbitmvi.orbit.compose.collectAsState
 import org.orbitmvi.orbit.compose.collectSideEffect
-import timber.log.Timber
 
 @Destination<ExternalModuleGraph>
 @Composable
@@ -125,78 +100,18 @@ fun PairAlertConditionScreen(
     fun getCurrentGroup() = state.pages.getOrNull(pagerState.currentPage)?.group
 
     viewModel.collectSideEffect { effect ->
-        when (effect) {
-            is PairAlertEffect.NavigateToAdd ->
-                navigator.navigate(
-                    AddPairAlertScreenDestination(
-                        pairAlertId = effect.pairId,
-                        groupId = getCurrentGroup()?.id,
-                    ),
-                )
-
-            PairAlertEffect.AskNotificationPermissionOnScreenOpen -> {
-                onScreenOpenNotificationPermissionLauncher
-                    .launch(Manifest.permission.POST_NOTIFICATIONS)
-            }
-
-            PairAlertEffect.AskNotificationPermissionOnNewPair -> {
-                onNewPairNotificationPermissionLauncher
-                    .launch(Manifest.permission.POST_NOTIFICATIONS)
-            }
-
-            is PairAlertEffect.SelectTab -> {
-                val page = state.pages.find { it.group.id == effect.groupId }!!
-                val pageIndex = state.pages.indexOf(page)
-                pagerState.scrollToPage(pageIndex)
-            }
-
-            is PairAlertEffect.ShowSnackbarAdded -> {
-                val pair = effect.pair
-                val aboveOrBelow =
-                    if (pair.above())
-                        ctx.getString(CoreRString.above)
-                    else
-                        ctx.getString(CoreRString.below)
-                val visuals =
-                    NotifyAddedSnackbarVisuals(
-                        title =
-                            ctx.getString(
-                                CoreRString.alert_snackbar_new_title,
-                                pair.targetCode,
-                            ),
-                        description =
-                            ctx.getString(
-                                CoreRString.alert_snackbar_new_desc,
-                                pair.targetCode,
-                                aboveOrBelow,
-                                CurrUtils.prepareToDisplay(pair.targetPrice),
-                                pair.baseCode,
-                            ),
-                    )
-                snackState.showSnackbar(visuals)
-            }
-
-            is PairAlertEffect.ShowRemovedSnackbar -> {
-                val visuals =
-                    NotifyRemovedSnackbarVisuals(
-                        title =
-                            ctx.getString(
-                                CoreRString.alert_snackbar_removed_title,
-                                effect.pair.targetCode,
-                            ),
-                        description =
-                            ctx.getString(
-                                CoreRString.alert_snackbar_removed_desc,
-                                effect.pair.targetCode,
-                                effect.pair.baseCode,
-                            ),
-                        onUndo = {
-                            viewModel.undoDelete(effect.pair)
-                        },
-                    )
-                snackState.showSnackbar(visuals)
-            }
-        }
+       handlePairAlertSideEffect(
+           effect,
+           state,
+           navigator,
+           viewModel,
+           pagerState,
+           snackState,
+           onScreenOpenNotificationPermissionLauncher,
+           onNewPairNotificationPermissionLauncher,
+           ctx,
+           ::getCurrentGroup,
+       )
     }
 
     Scaffold(
@@ -227,7 +142,7 @@ fun PairAlertConditionScreen(
         Box(modifier = Modifier.padding(it)) {
             when {
                 state.initialized.not() -> LoadingScreen()
-                isEmpty -> Empty(onNewPair = viewModel::onNewPair)
+                isEmpty -> PairAlertEmpty(onNewPair = viewModel::onNewPair)
                 else ->
                     Content(
                         component = component,
@@ -387,140 +302,6 @@ private fun GroupPage(
                     onDelete = { onDelete(it) },
                 )
                 AppHorDiv16()
-            }
-        }
-    }
-}
-
-@Composable
-private fun PairAlertItem(
-    component: PairAlertComponent,
-    pairAlert: PairAlert,
-    oneTimeTriggered: Boolean,
-    onClick: (PairAlert) -> Unit,
-    onEnableToggle: (PairAlert, Boolean) -> Unit,
-) {
-    var currencyName by remember {
-        mutableStateOf("")
-    }
-    val currencyRepo = component.currencyRepo()
-    LaunchedEffect(Unit) {
-        currencyName = currencyRepo.nameByCode(pairAlert.targetCode).name
-    }
-    Row(
-        modifier =
-            Modifier
-                .fillMaxWidth()
-                .background(Color.White)
-                .clickable {
-                    onClick(pairAlert)
-                }
-                .padding(horizontal = 24.dp, vertical = 16.dp),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        CurrIcon(modifier = Modifier.size(40.dp), code = pairAlert.targetCode)
-        Column(
-            modifier =
-                Modifier
-                    .padding(start = 12.dp)
-                    .weight(1f),
-            verticalArrangement = Arrangement.Center,
-        ) {
-            Text(
-                text =
-                    "$currencyName(${pairAlert.targetCode}) " +
-                        if (pairAlert.oneTimeNotRecurrent) "(One-time)" else "",
-                fontWeight = FontWeight.Medium,
-                color = ArkColor.TextPrimary,
-            )
-            Text(
-                text =
-                    buildString {
-                        append(
-                            "${
-                                if (pairAlert.above())
-                                    stringResource(CoreRString.above_c)
-                                else
-                                    stringResource(CoreRString.below_c)
-                            } ",
-                        )
-                        append("${CurrUtils.prepareToDisplay(pairAlert.targetPrice)} ")
-                        append(pairAlert.baseCode)
-                    },
-                color = ArkColor.TextTertiary,
-            )
-            if (oneTimeTriggered) {
-                val date = pairAlert.lastDateTriggered
-                date
-                    ?: Timber.e("Pair alert marked as triggered but lastDateTriggered is null")
-                if (date != null) {
-                    Text(
-                        text =
-                            stringResource(
-                                CoreRString.alert_notified_on,
-                                DateFormatUtils.notifiedOn(date),
-                            ),
-                        color = ArkColor.TextTertiary,
-                    )
-                }
-            }
-        }
-        Spacer(Modifier.width(8.dp))
-        Switch(
-            checked = pairAlert.enabled,
-            onCheckedChange = { onEnableToggle(pairAlert, it) },
-            colors =
-                SwitchDefaults.colors(
-                    checkedThumbColor = Color.White,
-                    checkedBorderColor = ArkColor.Primary,
-                    checkedTrackColor = ArkColor.Primary,
-                    uncheckedThumbColor = Color.White,
-                    uncheckedTrackColor = ArkColor.BGTertiary,
-                    uncheckedBorderColor = ArkColor.BGTertiary,
-                ),
-        )
-    }
-}
-
-@Composable
-private fun Empty(onNewPair: () -> Unit) {
-    Box(modifier = Modifier.fillMaxSize()) {
-        Column(
-            modifier = Modifier.align(Alignment.Center),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            Icon(
-                painter = painterResource(id = CoreRDrawable.ic_empty_pair),
-                contentDescription = "",
-                tint = Color.Unspecified,
-            )
-            Text(
-                modifier = Modifier.padding(top = 16.dp),
-                text = stringResource(CoreRString.alert_empty_title),
-                fontWeight = FontWeight.SemiBold,
-                fontSize = 20.sp,
-                color = ArkColor.TextPrimary,
-            )
-            Text(
-                modifier = Modifier.padding(top = 6.dp, start = 24.dp, end = 24.dp),
-                text = stringResource(CoreRString.alert_empty_desc),
-                fontSize = 14.sp,
-                lineHeight = 20.sp,
-                color = ArkColor.TextTertiary,
-                textAlign = TextAlign.Center,
-            )
-            AppButton(
-                modifier = Modifier.padding(top = 24.dp),
-                onClick = { onNewPair() },
-            ) {
-                Icon(
-                    painter = painterResource(id = CoreRDrawable.ic_add),
-                    contentDescription = "",
-                )
-                Text(
-                    modifier = Modifier.padding(start = 8.dp),
-                    text = stringResource(CoreRString.new_alert),
-                )
             }
         }
     }
