+++ feature/portfolio/src/main/java/dev/arkbuilders/rate/feature/portfolio/presentation/main/PortfolioScreen.kt
@@ -4,18 +4,13 @@ package dev.arkbuilders.rate.feature.portfolio.presentation.main
 
 import androidx.activity.compose.BackHandler
 import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.pager.PagerState
@@ -38,10 +33,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
@@ -54,20 +47,15 @@ import com.ramcosta.composedestinations.result.ResultRecipient
 import com.ramcosta.composedestinations.result.onResult
 import dev.arkbuilders.rate.core.domain.CurrUtils
 import dev.arkbuilders.rate.core.domain.model.CurrencyCode
-import dev.arkbuilders.rate.core.presentation.CoreRDrawable
 import dev.arkbuilders.rate.core.presentation.CoreRString
 import dev.arkbuilders.rate.core.presentation.theme.ArkColor
-import dev.arkbuilders.rate.core.presentation.ui.AppButton
 import dev.arkbuilders.rate.core.presentation.ui.AppHorDiv16
 import dev.arkbuilders.rate.core.presentation.ui.AppSwipeToDismiss
-import dev.arkbuilders.rate.core.presentation.ui.CurrIcon
 import dev.arkbuilders.rate.core.presentation.ui.GroupViewPager
 import dev.arkbuilders.rate.core.presentation.ui.LargeNumberText
 import dev.arkbuilders.rate.core.presentation.ui.LargeNumberTooltipBox
 import dev.arkbuilders.rate.core.presentation.ui.LoadingScreen
 import dev.arkbuilders.rate.core.presentation.ui.NoResult
-import dev.arkbuilders.rate.core.presentation.ui.NotifyAddedSnackbarVisuals
-import dev.arkbuilders.rate.core.presentation.ui.NotifyRemovedSnackbarVisuals
 import dev.arkbuilders.rate.core.presentation.ui.RateSnackbarHost
 import dev.arkbuilders.rate.core.presentation.ui.SearchTextField
 import dev.arkbuilders.rate.core.presentation.ui.group.EditGroupOptionsBottomSheet
@@ -117,52 +105,15 @@ fun PortfolioScreen(
     fun getCurrentGroup() = state.pages.getOrNull(pagerState.currentPage)?.group
 
     viewModel.collectSideEffect { effect ->
-        when (effect) {
-            is PortfolioScreenEffect.SelectTab -> {
-                val page = state.pages.find { it.group.id == effect.groupId }!!
-                val pageIndex = state.pages.indexOf(page)
-                pagerState.scrollToPage(pageIndex)
-            }
-
-            is PortfolioScreenEffect.ShowSnackbarAdded -> {
-                val added =
-                    effect.assets
-                        .joinToString {
-                            "${CurrUtils.prepareToDisplay(BigDecimal(it.value))} ${it.code}"
-                        }
-                val visuals =
-                    NotifyAddedSnackbarVisuals(
-                        ctx.getString(CoreRString.portfolio_snackbar_new_title),
-                        ctx.getString(
-                            CoreRString.portfolio_snackbar_new_desc,
-                            added,
-                        ),
-                    )
-
-                snackState.showSnackbar(visuals)
-            }
-
-            is PortfolioScreenEffect.ShowRemovedSnackbar -> {
-                val removed =
-                    CurrUtils.prepareToDisplay(effect.asset.value) +
-                        " ${effect.asset.code}"
-                val visuals =
-                    NotifyRemovedSnackbarVisuals(
-                        title = ctx.getString(CoreRString.portfolio_snackbar_removed_title),
-                        description =
-                            ctx.getString(
-                                CoreRString.portfolio_snackbar_removed_desc,
-                                removed,
-                            ),
-                        onUndo = {
-                            viewModel.undoDelete(effect.asset)
-                        },
-                    )
-                snackState.showSnackbar(visuals)
-            }
-
-            PortfolioScreenEffect.NavigateBack -> navigator.popBackStack()
-        }
+        handlePortfolioSideEffect(
+            effect,
+            viewModel,
+            navigator,
+            state,
+            pagerState,
+            snackState,
+            ctx,
+        )
     }
 
     Scaffold(
@@ -373,124 +324,5 @@ private fun GroupPage(
         }
     } else {
         NoResult()
-    }
-}
-
-@Composable
-private fun CurrencyItem(
-    amount: PortfolioDisplayAsset,
-    onClick: (PortfolioDisplayAsset) -> Unit,
-) {
-    Row(
-        modifier =
-            Modifier
-                .fillMaxWidth()
-                .background(Color.White)
-                .clickable {
-                    onClick(amount)
-                }
-                .padding(horizontal = 24.dp, vertical = 16.dp),
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        CurrIcon(modifier = Modifier.size(40.dp), code = amount.asset.code)
-        Column(
-            modifier = Modifier.padding(start = 12.dp),
-            verticalArrangement = Arrangement.Center,
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-            ) {
-                Text(
-                    modifier = Modifier.padding(end = 8.dp),
-                    text = amount.asset.code,
-                    fontWeight = FontWeight.Medium,
-                    color = ArkColor.TextPrimary,
-                )
-                LargeNumberTooltipBox(
-                    modifier = Modifier.weight(1f),
-                    number = amount.baseAmount.value,
-                    code = amount.baseAmount.code,
-                ) {
-                    LargeNumberText(
-                        number = amount.baseAmount.value,
-                        code = amount.baseAmount.code,
-                        fontWeight = FontWeight.Medium,
-                        color = ArkColor.TextPrimary,
-                        fontSize = 14.sp,
-                        textAlign = TextAlign.End,
-                    )
-                }
-            }
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-            ) {
-                Text(
-                    modifier = Modifier.padding(end = 8.dp),
-                    text = CurrUtils.prepareToDisplay(amount.ratioToBase),
-                    color = ArkColor.TextTertiary,
-                )
-                LargeNumberTooltipBox(
-                    modifier = Modifier.weight(1f),
-                    number = amount.asset.value,
-                    code = amount.asset.code,
-                ) {
-                    LargeNumberText(
-                        number = amount.asset.value,
-                        code = amount.asset.code,
-                        color = ArkColor.TextTertiary,
-                        fontSize = 14.sp,
-                        textAlign = TextAlign.End,
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun PortfolioEmpty(navigator: DestinationsNavigator) {
-    Box(modifier = Modifier.fillMaxSize()) {
-        Column(
-            modifier = Modifier.align(Alignment.Center),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            Icon(
-                painter = painterResource(id = CoreRDrawable.ic_empty_portfolio),
-                contentDescription = "",
-                tint = Color.Unspecified,
-            )
-            Text(
-                modifier = Modifier.padding(top = 16.dp),
-                text = stringResource(CoreRString.portfolio_empty_title),
-                fontWeight = FontWeight.SemiBold,
-                fontSize = 20.sp,
-                color = ArkColor.TextPrimary,
-            )
-            Text(
-                modifier = Modifier.padding(top = 6.dp, start = 24.dp, end = 24.dp),
-                text = stringResource(CoreRString.portfolio_empty_desc),
-                fontSize = 14.sp,
-                lineHeight = 20.sp,
-                color = ArkColor.TextTertiary,
-                textAlign = TextAlign.Center,
-            )
-            AppButton(
-                modifier = Modifier.padding(top = 24.dp),
-                onClick = {
-                    navigator.navigate(AddAssetScreenDestination())
-                },
-            ) {
-                Icon(
-                    painter = painterResource(id = CoreRDrawable.ic_add),
-                    contentDescription = "",
-                )
-                Text(
-                    modifier = Modifier.padding(start = 8.dp),
-                    text = stringResource(CoreRString.portfolio_empty_new_assets),
-                )
-            }
-        }
     }
 }
