+++ feature/quick/src/main/java/dev/arkbuilders/rate/feature/quick/presentation/main/QuickScreen.kt
@@ -1,20 +1,12 @@
 package dev.arkbuilders.rate.feature.quick.presentation.main
 
 import androidx.activity.compose.BackHandler
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.pager.PagerState
@@ -27,28 +19,17 @@ import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHostState
-import androidx.compose.material3.Text
 import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import androidx.constraintlayout.compose.ConstraintLayout
-import androidx.constraintlayout.compose.Dimension
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.annotation.ExternalModuleGraph
@@ -56,17 +37,12 @@ import com.ramcosta.composedestinations.generated.quick.destinations.AddQuickScr
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import com.ramcosta.composedestinations.result.ResultRecipient
 import com.ramcosta.composedestinations.result.onResult
-import dev.arkbuilders.rate.core.domain.CurrUtils
 import dev.arkbuilders.rate.core.domain.model.Amount
 import dev.arkbuilders.rate.core.domain.model.CurrencyCode
 import dev.arkbuilders.rate.core.domain.model.CurrencyName
-import dev.arkbuilders.rate.core.presentation.CoreRDrawable
 import dev.arkbuilders.rate.core.presentation.CoreRString
-import dev.arkbuilders.rate.core.presentation.R
 import dev.arkbuilders.rate.core.presentation.theme.ArkColor
-import dev.arkbuilders.rate.core.presentation.ui.AppButton
 import dev.arkbuilders.rate.core.presentation.ui.AppHorDiv16
-import dev.arkbuilders.rate.core.presentation.ui.CurrIcon
 import dev.arkbuilders.rate.core.presentation.ui.CurrencyInfoItem
 import dev.arkbuilders.rate.core.presentation.ui.GroupViewPager
 import dev.arkbuilders.rate.core.presentation.ui.ListHeader
@@ -81,7 +57,6 @@ import dev.arkbuilders.rate.core.presentation.ui.group.EditGroupRenameBottomShee
 import dev.arkbuilders.rate.core.presentation.ui.group.EditGroupReorderBottomSheet
 import dev.arkbuilders.rate.core.presentation.ui.group.EditGroupRow
 import dev.arkbuilders.rate.core.presentation.utils.DateFormatUtils
-import dev.arkbuilders.rate.core.presentation.utils.findActivity
 import dev.arkbuilders.rate.feature.quick.di.QuickComponentHolder
 import dev.arkbuilders.rate.feature.quick.domain.model.PinnedQuickPair
 import dev.arkbuilders.rate.feature.quick.domain.model.QuickPair
@@ -123,56 +98,14 @@ fun QuickScreen(
     val pagerState = rememberPagerState { state.pages.size }
     val snackState = remember { SnackbarHostState() }
     viewModel.collectSideEffect { effect ->
-        when (effect) {
-            is QuickScreenEffect.ShowSnackbarAdded -> {
-                val added =
-                    ctx.getString(
-                        R.string.quick_snackbar_new_added_to,
-                        effect.pair.from,
-                        effect.pair.to.joinToString { it.code },
-                    )
-                val visuals =
-                    NotifyAddedSnackbarVisuals(
-                        title = ctx.getString(R.string.quick_snackbar_new_title),
-                        description =
-                            ctx.getString(
-                                R.string.quick_snackbar_new_desc,
-                                added,
-                            ),
-                    )
-                snackState.showSnackbar(visuals)
-            }
-
-            is QuickScreenEffect.ShowRemovedSnackbar -> {
-                val removed =
-                    ctx.getString(
-                        CoreRString.quick_snackbar_new_added_to,
-                        effect.pair.from,
-                        effect.pair.to.joinToString { it.code },
-                    )
-                val visuals =
-                    NotifyRemovedSnackbarVisuals(
-                        title = ctx.getString(CoreRString.quick_snackbar_removed_title),
-                        description =
-                            ctx.getString(
-                                CoreRString.quick_snackbar_removed_desc,
-                                removed,
-                            ),
-                        onUndo = {
-                            viewModel.undoDelete(effect.pair)
-                        },
-                    )
-                snackState.showSnackbar(visuals)
-            }
-
-            is QuickScreenEffect.SelectTab -> {
-                val page = state.pages.find { it.group.id == effect.groupId }!!
-                val pageIndex = state.pages.indexOf(page)
-                pagerState.scrollToPage(pageIndex)
-            }
-
-            QuickScreenEffect.NavigateBack -> ctx.findActivity()?.finish()
-        }
+        handleQuickSideEffect(
+            effect,
+            state,
+            viewModel,
+            pagerState,
+            snackState,
+            ctx,
+        )
     }
 
     val isEmpty = state.pages.isEmpty()
@@ -348,7 +281,7 @@ private fun Content(
             onFilterChanged(it)
         }
         if (state.filter.isNotEmpty()) {
-            SearchPage(
+            QuickSearchPage(
                 filter = state.filter,
                 topResults = state.topResults,
                 onNewCode = onNewCode,
@@ -475,195 +408,6 @@ private fun GroupPage(
     }
 }
 
-@Composable
-private fun SearchPage(
-    filter: String,
-    topResults: List<CurrencyName>,
-    onNewCode: (CurrencyCode) -> Unit,
-) {
-    val filtered =
-        topResults.filter {
-            it.name.contains(filter, ignoreCase = true) ||
-                it.code.contains(filter, ignoreCase = true)
-        }
-    if (filtered.isNotEmpty()) {
-        LazyColumn(modifier = Modifier.fillMaxSize()) {
-            item {
-                ListHeader(text = stringResource(CoreRString.top_results))
-            }
-            items(filtered) { name ->
-                CurrencyInfoItem(name) { onNewCode(it.code) }
-            }
-        }
-    } else {
-        NoResult()
-    }
-}
-
-@Composable
-private fun QuickItem(
-    from: Amount,
-    to: List<Amount>,
-    dateText: String,
-    onClick: () -> Unit,
-) {
-    var expanded by remember {
-        mutableStateOf(false)
-    }
-
-    ConstraintLayout(
-        modifier =
-            Modifier
-                .fillMaxWidth()
-                .background(Color.White)
-                .clickable {
-                    onClick()
-                },
-    ) {
-        val (icons, content, chevron) = createRefs()
-        Row(
-            modifier =
-                Modifier.constrainAs(icons) {
-                    top.linkTo(parent.top, margin = 16.dp)
-                    start.linkTo(parent.start, margin = 24.dp)
-                },
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Box(
-                modifier =
-                    Modifier
-                        .size(40.dp),
-            ) {
-                CurrIcon(modifier = Modifier.size(40.dp), code = from.code)
-            }
-            if (!expanded) {
-                Box(
-                    modifier =
-                        Modifier
-                            .size(40.dp)
-                            .offset((-12).dp)
-                            .border(2.dp, Color.White, CircleShape),
-                ) {
-                    if (to.size == 1) {
-                        CurrIcon(
-                            modifier =
-                                Modifier
-                                    .size(38.dp)
-                                    .align(Alignment.Center)
-                                    .clip(CircleShape)
-                                    .background(Color.White),
-                            code = to.first().code,
-                        )
-                    } else {
-                        Box(
-                            modifier =
-                                Modifier
-                                    .size(40.dp)
-                                    .background(ArkColor.BGTertiary, CircleShape),
-                        ) {
-                            Text(
-                                modifier = Modifier.align(Alignment.Center),
-                                text = "+ ${to.size}",
-                                fontWeight = FontWeight.SemiBold,
-                                fontSize = 16.sp,
-                                color = ArkColor.TextTertiary,
-                            )
-                        }
-                    }
-                }
-            }
-        }
-        Column(
-            modifier =
-                Modifier
-                    .constrainAs(content) {
-                        start.linkTo(icons.end)
-                        if (to.size > 1)
-                            end.linkTo(chevron.start)
-                        else
-                            end.linkTo(parent.end, margin = 24.dp)
-                        top.linkTo(parent.top, margin = 16.dp)
-                        bottom.linkTo(parent.bottom, margin = 16.dp)
-                        width = Dimension.fillToConstraints
-                    }
-                    .padding(start = if (expanded) 12.dp else 0.dp),
-            verticalArrangement = Arrangement.Center,
-        ) {
-            Text(
-                text =
-                    "${from.code} to " +
-                        to.joinToString(", ") { it.code },
-                fontWeight = FontWeight.Medium,
-                color = ArkColor.TextPrimary,
-            )
-            if (expanded) {
-                Text(
-                    text = "${CurrUtils.prepareToDisplay(from.value)} ${from.code} =",
-                    color = ArkColor.TextTertiary,
-                )
-                to.forEach {
-                    Row(
-                        modifier = Modifier.padding(top = 8.dp),
-                        verticalAlignment = Alignment.CenterVertically,
-                    ) {
-                        CurrIcon(modifier = Modifier.size(20.dp), code = it.code)
-                        Text(
-                            modifier = Modifier.padding(start = 8.dp),
-                            text = "${CurrUtils.prepareToDisplay(it.value)} ${it.code}",
-                            color = ArkColor.TextTertiary,
-                        )
-                    }
-                }
-            } else {
-                Text(
-                    text =
-                        "${CurrUtils.prepareToDisplay(from.value)} ${from.code} = " +
-                            "${CurrUtils.prepareToDisplay(to.first().value)} ${to.first().code}",
-                    color = ArkColor.TextTertiary,
-                )
-            }
-            Text(
-                modifier = Modifier.padding(top = if (expanded) 8.dp else 0.dp),
-                text = dateText,
-                color = ArkColor.TextTertiary,
-                fontSize = 12.sp,
-            )
-        }
-        if (to.size > 1) {
-            Box(
-                modifier =
-                    Modifier
-                        .constrainAs(chevron) {
-                            height = Dimension.fillToConstraints
-                            top.linkTo(parent.top)
-                            end.linkTo(parent.end)
-                            bottom.linkTo(parent.bottom)
-                        }
-                        .clickable {
-                            expanded = !expanded
-                        }
-                        .padding(start = 13.dp, end = 29.dp, top = 23.dp),
-            ) {
-                if (expanded) {
-                    Icon(
-                        modifier = Modifier,
-                        painter = painterResource(CoreRDrawable.ic_chevron_up),
-                        contentDescription = "",
-                        tint = ArkColor.FGSecondary,
-                    )
-                } else {
-                    Icon(
-                        modifier = Modifier,
-                        painter = painterResource(CoreRDrawable.ic_chevron),
-                        contentDescription = "",
-                        tint = ArkColor.FGSecondary,
-                    )
-                }
-            }
-        }
-    }
-}
-
 @Preview(showBackground = true)
 @Composable
 private fun PreviewItem() {
@@ -673,50 +417,4 @@ private fun PreviewItem() {
         dateText = "Calculated on",
         onClick = {},
     )
-}
-
-@Composable
-private fun QuickEmpty(navigator: DestinationsNavigator) {
-    Box(modifier = Modifier.fillMaxSize()) {
-        Column(
-            modifier = Modifier.align(Alignment.Center),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            Icon(
-                painter = painterResource(id = CoreRDrawable.ic_empty_quick),
-                contentDescription = "",
-                tint = Color.Unspecified,
-            )
-            Text(
-                modifier = Modifier.padding(top = 16.dp),
-                text = stringResource(CoreRString.quick_empty_title),
-                fontWeight = FontWeight.SemiBold,
-                fontSize = 20.sp,
-                color = ArkColor.TextPrimary,
-            )
-            Text(
-                modifier = Modifier.padding(top = 6.dp, start = 24.dp, end = 24.dp),
-                text = stringResource(CoreRString.quick_empty_desc),
-                fontSize = 14.sp,
-                lineHeight = 20.sp,
-                color = ArkColor.TextTertiary,
-                textAlign = TextAlign.Center,
-            )
-            AppButton(
-                modifier = Modifier.padding(top = 24.dp),
-                onClick = {
-                    navigator.navigate(AddQuickScreenDestination())
-                },
-            ) {
-                Icon(
-                    painter = painterResource(id = CoreRDrawable.ic_add),
-                    contentDescription = "",
-                )
-                Text(
-                    modifier = Modifier.padding(start = 8.dp),
-                    text = stringResource(CoreRString.calculate),
-                )
-            }
-        }
-    }
 }
